Manual fix for the type here.
Even though the headers "correctly" declare it as a void pointer,
the actual usage is as a ptrdiff_t in (video card) memory, so we
have to interpret the number as integer. The functions of this type
will need manual care now.

Likely, this means that the XS autogeneration needs to automatically
exclude functions that have been implemented/written manually in Glew.xs
or the generation needs to be made smarter about special functions
where a void* means "use an integer" as opposed to my usual usage
of "use the string buffer".