glNamedBufferData() isn't supported on Win10+Intel
It would be nice to write abstractions/reimplementations of these
but that means chasing more yaks to shave. Such an abstraction/wrapper
would need to do the three steps of saving the current buffer,
rebinding the next buffer, setting the buffer data and then rebinding
the old buffer. No fun.